Streamlit/requirements.txt
streamlit>=1.34
nbformat>=5.10
nbclient>=0.10
gitpython>=3.1
openai>=1.40
pandas>=2.2
numpy>=1.26
tqdm
requests
python-dateutil
